requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dash==3.3.0
Flask==3.1.2
fonttools==4.61.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.14.0
nest-asyncio==1.6.0
numpy==2.4.0
packaging==25.0
pillow==12.0.0
plotly==6.5.0
pyparsing==3.3.1
pyserial==3.5
python-dateutil==2.9.0.post0
requests==2.32.5
retrying==1.4.2
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
zipp==3.23.0
